Fjern gammelt gui widget system.
