--- core/src/main/java/ml/volder/unikapi/widgets/RenderRelative.java
@@ -1,8 +1,0 @@
-package ml.volder.unikapi.widgets;
-
-public enum RenderRelative {
-    LEFT_TOP,
-    LEFT_BOTTOM,
-    RIGHT_TOP,
-    RIGHT_BOTTOM
-}
